requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-heroku==0.3.1
django-rest-knox==3.1.5
djangorestframework==3.8.2
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
psycopg2==2.7.5